SolrSchemaPreservingSolrRDD initial refactoring
